[synthetics] Move non-exported utils to `batch.ts`
